melhorando rota de remoção de User e refatorando código

+++ src/main/java/domain/user/UserService.java
@@ -16,33 +16,38 @@ import javax.inject.Inject;
  */
 @ApplicationScoped
 public class UserService {
-    
+
     @Inject
     private UserDB userDB;
-    
-    
-    public void save(User user){
+
+    public void save(User user) {
         User userCpf = this.userDB.getUserByCpf(user.getCpf());
         User userEmail = this.userDB.getUserByEmail(user.getEmail());
-        if(userCpf !=null || userEmail !=null){
-            throw new BusinessException("CPF ou Email não está disponível para ser cadastrado.");          
+        if (userCpf != null || userEmail != null) {
+            throw new BusinessException("CPF ou Email não está disponível para ser cadastrado.");
         }
         this.userDB.save(user);
     }
-    
-    public void update(User user){
+
+    public void update(User user) {
         this.userDB.update(user);
     }
-    
-    public User getUserByCpf(String cpf){
+
+    public User getUserByCpf(String cpf) {
         return this.userDB.getUserByCpf(cpf);
     }
-    
-    public User getUserByEmail(String email){
+
+    public User getUserByEmail(String email) {
         return this.userDB.getUserByEmail(email);
     }
-    
-    public void remove(User user){
-        this.userDB.remove(user);
+
+    public void remove(User user) {
+        User userToValidatePassword = this.getUserByCpf(user.getCpf());
+        if (userToValidatePassword.getPassword().equals(user.getPassword())) {
+            this.userDB.remove(user);
+        } else {
+            throw new BusinessException("Senha incorreta");
+        }
+
     }
 }
